refactor(elements): extract Resolve+Navigation+Observation+Cookies+PhaseControls+Assembly+ActionMediator sub-modules
Phase 12a §8 of the CreateElementMediator god-module split — the FINAL
landing. Trims the façade from 915 LoC to ~38 LoC (≤150 LoC target)
by relocating seven cohesive clusters into sibling sub-modules under
Create/. After this commit the façade contains only:

  - The public `createElementMediator(page)` factory (≤10 LoC body)
  - A pass-through `extractActionMediator` re-export
  - A pass-through `getActivePhase` / `getActiveStage` re-export

…and a single barrel import. End-state of the Phase 12a contract.

Why
---
After §7 (FieldResolve+Wait+Discover) the parent still owned the
ENTIRE click-resolution, navigation, observation, cookie, phase-
control, mediator-assembly, and action-mediator surfaces — eight
distinct cognitive clusters glued together with 31 helper functions
and 10 type aliases. None of them depend on each other except via
narrow bundle types, which makes the seven-way split mechanical:
each cluster is a self-contained vertical with its own constants,
private helpers, and `buildXxxCluster` aggregator.

Bundling all seven into a single commit (instead of spreading them
across seven commits) is justified because:
  - Each individual sub-module is small (40-330 LoC).
  - Splitting them would front-load the husky 23-gate run (~30 min
    each) seven times for the same net code motion.
  - The public-surface invariant is identical for every sub-step
    (byte-identical d.ts/d.cts), so testing them together gives
    the same confidence in 1/7th the wall-time.

What
----
- Create/Resolve.ts (new, ~340 LoC) — owns all resolveVisible* impls,
  resolveAndClick + force-click fallback, the four resolver-method
  builders (buildResolveVisible/buildResolveAllVisible/
  buildResolveVisibleInContext/buildResolveAndClick), buildGetFormAnchor,
  ResolveBundle + FormBundle types, and the two cluster aggregators
  (buildResolveCluster, buildFormCluster).
- Create/Navigation.ts (new, ~140 LoC) — owns NETWORK_IDLE_TIMEOUT,
  URL_WAIT_TIMEOUT, the five nav method builders, NavBundle type,
  and buildNavCluster (preserves single-source waitForNetworkIdle).
- Create/Observation.ts (new, ~140 LoC) — owns extractRawHrefs +
  the four read-only DOM probe builders (getAttributeValue,
  checkAttribute, countByText, countBySelector, collectAllHrefs),
  AttrBundle + CountBundle types, and buildAttrCluster +
  buildCountCluster aggregators.
- Create/Cookies.ts (new, ~55 LoC) — owns buildGetCookies +
  buildAddCookies, CookieBundle type, and buildCookieCluster.
- Create/PhaseControls.ts (new, ~70 LoC) — owns buildPhaseControls
  + buildStaticCluster + the merged PhaseControlsBundle/StaticBundle
  types. Imports setActivePhase/setActiveStage DIRECTLY from
  ActiveState (no longer routed through the Create/index.ts
  transient pass-through, which is now dead). Re-exports
  getActivePhase/getActiveStage so the façade's public re-export
  has a single owning module.
- Create/Assembly.ts (new, ~50 LoC) — owns assembleElementMediator,
  the function that composes all six bundle builders into the full
  Omit<IElementMediator, 'network'> shape.
- Create/ActionMediator.ts (new, ~280 LoC) — owns snapshotSessionStorage,
  the four action-side bundle types (FrameActionBundle, ActionPassThroughBundle,
  ActionStorageBundle, ActionNetworkBundle), 12 buildXxx helpers, and
  the public extractActionMediator factory.
- Create/index.ts — barrel re-exports the seven new sub-module surfaces;
  drops the ActiveState transient pass-through (setActivePhase /
  setActiveStage no longer surface through the barrel; getActivePhase /
  getActiveStage now route through PhaseControls).
- CreateElementMediator.ts — slashed from 915 LoC to 38 LoC. Pulls in
  4 import sources (down from 11): playwright-core (type Page),
  ../Network/NetworkDiscovery.js (createNetworkDiscovery), the
  Create/index.js barrel (assembleElementMediator,
  extractActionMediator, getActivePhase, getActiveStage, IFormCache),
  and ./ElementMediator.js (IElementMediator).
- .github/docs-coverage-allowlist.txt — 17 allowlist entries (Reason
  per symbol) for the new internal exports.

Public surface
--------------
lib/index.d.ts + lib/index.d.cts SHA256 BYTE-IDENTICAL to baseline
(both 14823 B, hash 221E1611…AAB055).
lib/index.cjs grew by ~3 KB for the seven new module init wrappers —
implementation reorganisation, not API change. Public scraper API
is unchanged: createElementMediator, extractActionMediator,
getActivePhase, getActiveStage continue to surface from
CreateElementMediator.ts with identical signatures and identical
function-identity preservation through the assemble spread chain.

Verification
------------
- tsc --noEmit: clean
- eslint Mediator/Elements/Create + parent: clean (2 prefer-default-export
  hits resolved by adding type re-exports in Assembly.ts + ActionMediator.ts)
- prettier: clean (Resolve.ts auto-formatted)
- lint:architecture (src/Scrapers/Pipeline): 0 violations
- jest Mediator/Elements (30 suites / 270 tests): all pass
- npm run build: success; d.ts hashes match baseline (BYTE-IDENTICAL)

End-state
---------
Phase 12a is complete: CreateElementMediator.ts now a 38-LoC public
façade. The 10 Create/ sub-modules are all ≤330 LoC, all under the
import-x/max-dependencies cap of 15, all conformant with the
custom architecture lint (Rule #15 + others). 270/270 Mediator/Elements
tests pass. Public surface preserved byte-identical end-to-end.
